Create functions.ts and Shift functions and documentation there.
